refactor: split heavy media helpers out of utils into media_utils
